🎨 Created component for shoppinglist cards
